Se corrigieron cotas de PistaGrafico
Se agregó un chequeo de la longitud de la cota, si es muy corta las
flechas se grafican en el exterior de la cota apuntando hacia adentro.
También se corrigió un caso borde en el que el que el bounding rect no
cubría el número de la cota correctamente. Se actualizó el gráfico svg
para mostrar estos casos borde.

+++ interfaz/aeropuerto/dibujo/cotagrafica.h
@@ -39,15 +39,15 @@ private:
     QPointF punto1;
     QPointF punto2;
     Sentido sentido;
-    float distanciaPerp; //Distancia perpendicular al eje entre los dos puntos en que se encuentra la linea de cota
-    float longitud;      //Longitud entre los puntos
+    qreal distanciaPerp; //Distancia perpendicular al eje entre los dos puntos en que se encuentra la linea de cota
+    qreal longitud;      //Longitud entre los puntos
     QFont fuente;
     qreal margin;
     qreal tramoAdicional;
 
     QRectF bRect;
     QPainterPath pPath;
-    float anchoRect;
+    qreal anchoRect;
 
     QString texto;
     qreal anchoTexto;
